refactor: normalize ChatUserMessage component
