+++ web/.storybook/stories-category-doc-pages/Chat.stories.mdx
@@ -1,6 +1,5 @@
 import { Meta } from '@storybook/addon-docs';
 import { Typography } from 'antd';
-import UserChatMessage from '../../components/chat/ChatUserMessage';
 import { ChatMessage } from '../../interfaces/chat-message.model';
 
 <Meta title="Owncast/Documentation/Chat" />
