--- web/components/chat/ChatUserMessage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ChatUserMessage';
