--- web/components/chat/index.ts
@@ -1,1 +1,0 @@
-export { default as ChatUserMessage } from './ChatUserMessage';
